Woo: Move product variations to top of sites state
The previous code had product variations underneath the products
directory as they depend on the products, but from a reducer standpoint,
the products reducer isn't a combined reducer, so it's not designed to
work this way. So this code puts variations underneath sites for now.
Perhaps in the future, it could be put under products at some point if
it were refactored to support it.
