+++ client/extensions/woocommerce/state/sites/product-variations/reducer.js
@@ -3,6 +3,18 @@
  */
 import { isEqual } from 'lodash';
 
+/**
+ * Internal dependencies
+ */
+import { createReducer } from 'state/utils';
+import {
+	WOOCOMMERCE_PRODUCT_VARIATION_UPDATED,
+} from 'woocommerce/state/action-types';
+
+export default createReducer( {}, {
+	[ WOOCOMMERCE_PRODUCT_VARIATION_UPDATED ]: variationUpdated,
+} );
+
 export function variationUpdated( state, action ) {
 	const { data } = action;
 	const variations = state && state.variations || [];
